Selection Items Hospital Ward ID

+++ prisma/migrations/20260723120000_add_t_hospital_ward/migration.sql
@@ -1,6 +1,6 @@
 -- CreateTable
 CREATE TABLE "t_hospital_ward" (
-    "code" DECIMAL(2,0) NOT NULL,
+    "code" SERIAL NOT NULL,
     "ward" VARCHAR(50) NOT NULL,
 
     CONSTRAINT "t_hospital_ward_pkey" PRIMARY KEY ("code")
